refactor client controller
