+++ bsn/rollup-finality-provider/clientcontroller/consumer.go
@@ -1,4 +1,4 @@
-package opstackl2
+package clientcontroller
 
 import (
 	"context"
